refactor(rql): move logical planning + optimizer into reddb-io-rql
Move the storage-agnostic half of query planning into reddb-io-rql,
completing the crate contract: text in -> typed logical plan out
(ADR 0053). The multi-pass AST rewriter, the pluggable optimization-pass
pipeline, projection routing, pathkey/sort-order reasoning, and the
self-contained filter-ranking / decorrelation / stats-collection models
consume no storage statistics, index metadata, or executor capabilities,
so they depend only on the canonical AST, its SQL lowering helpers, and
reddb-io-types. Cost estimation, histograms, stats catalog/provider,
partition/hypertable pruning, join-order DP, the plan cache, and the
catalog-bound CanonicalPlanner stay in reddb-server and consume the
crate's output. Tests move verbatim.

Refs #1106
