+++ crates/reddb-rql/src/lib.rs
@@ -47,7 +47,9 @@ pub mod filter_optimizer;
 pub mod lexer;
 pub mod limits;
 pub mod modes;
+pub mod optimizer;
 pub mod parser;
+pub mod planner;
 pub mod renderer;
 pub mod sql;
 pub mod sql_lowering;
