+++ crates/reddb-rql/src/optimizer/stats.rs
