+++ crates/reddb-rql/src/planner/projections.rs
